Updated Relational mapping
