+++ tenttalk-api/src/main/java/org/codestars/tenttalk_api/service/ReviewService.java
@@ -1,15 +1,49 @@
 package org.codestars.tenttalk_api.service;
 
 import org.codestars.tenttalk_api.dto.ReviewDTO;
+import org.codestars.tenttalk_api.models.Campground;
+import org.codestars.tenttalk_api.models.Review;
+import org.codestars.tenttalk_api.models.data.CampgroundRepository;
+import org.codestars.tenttalk_api.models.data.ReviewRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+@Service
+public class ReviewService {
 
-public interface ReviewService {
+    @Autowired
+    ReviewRepository reviewRepository;
+    @Autowired
+    CampgroundRepository campgroundRepository;
 
-    ReviewDTO saveReview(ReviewDTO reviewDTO);
-    List<ReviewDTO>getAllReviews();
-    ReviewDTO getReviewById(Long id);
-    ReviewDTO updateReview(Long id, ReviewDTO reviewDTO);
-    void deleteReview(Long id);
+    @Transactional
+    public Review addReview(ReviewDTO reviewDTO) {
+
+        Campground campground = campgroundRepository.findById(reviewDTO.getCampgroundId())
+                .orElseThrow(() -> new RuntimeException("Campground not found"));
+
+        Review review = new Review();
+        review.setFeedback(reviewDTO.getFeedback());
+        review.setRating(reviewDTO.getRating());
+        review.setCampground(campground);
+        return reviewRepository.save(review);
+    }
+
+    @Transactional
+    public Review updateReview (Long id, ReviewDTO reviewDTO){
+        Review review = reviewRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Review not found"));
+
+        review.setFeedback((reviewDTO.getFeedback()));
+        review.setRating(reviewDTO.getRating());
+
+        if (reviewDTO.getCampgroundId() != null) {
+            Campground campground = campgroundRepository.findById(reviewDTO.getCampgroundId())
+                    .orElseThrow(() -> new RuntimeException("Campground not found"));
+            review.setCampground(campground);
+        }
+        return reviewRepository.save(review);
+    }
 
 }
